refactor: Move files to api folder, change paths
