+++ server/api/categories.js
@@ -1,26 +1,26 @@
 const clone = require('clone')
-const config = require('./config')
+const config = require('../config')
 
 let db = {}
 
 const defaultData = {
   categories: [
-      {
-        name: 'react',
-        path: 'react'
-      },
-      {
-        name: 'redux',
-        path: 'redux'
-      },
-      {
-        name: 'udacity',
-        path: 'udacity'
-      }
+    {
+      name: 'react',
+      path: 'react'
+    },
+    {
+      name: 'redux',
+      path: 'redux'
+    },
+    {
+      name: 'udacity',
+      path: 'udacity'
+    }
   ]
 }
 
-function getData (token) {
+function getData(token) {
   //Each token has it's own copy of the DB. The token in this case is like an app id.
   let data = db[token]
   //This populates the default user data if there isn't any in the db.
@@ -30,9 +30,9 @@ function getData (token) {
   return data
 }
 
-function getAll (token) {
+function getAll(token) {
   return new Promise((res) => {
-    res(getData(token))    
+    res(getData(token))
   })
 }
