+++ server/api/comments.js
@@ -11,7 +11,7 @@ const defaultData = {
     author: 'thingtwo',
     voteScore: 6,
     deleted: false,
-    parentDeleted: false 
+    parentDeleted: false
   },
   "8tu4bsun805n8un48ve89": {
     id: '8tu4bsun805n8un48ve89',
@@ -25,7 +25,7 @@ const defaultData = {
   }
 }
 
-function getData (token) {
+function getData(token) {
   let data = db[token]
   if (data == null) {
     data = db[token] = clone(defaultData)
@@ -33,7 +33,7 @@ function getData (token) {
   return data
 }
 
-function getByParent (token, parentId) {
+function getByParent(token, parentId) {
   return new Promise((res) => {
     let comments = getData(token)
     let keys = Object.keys(comments)
@@ -42,18 +42,18 @@ function getByParent (token, parentId) {
   })
 }
 
-function get (token, id) {
+function get(token, id) {
   return new Promise((res) => {
     const comments = getData(token)
     res(
       comments[id].deleted || comments[id].parentDeleted
         ? {}
-        : comments[id]      
-      )
+        : comments[id]
+    )
   })
 }
 
-function add (token, comment) {
+function add(token, comment) {
   return new Promise((res) => {
     let comments = getData(token)
 
@@ -67,55 +67,55 @@ function add (token, comment) {
       deleted: false,
       parentDeleted: false
     }
-     
+
     res(comments[comment.id])
   })
 }
 
-function vote (token, id, option) {
+function vote(token, id, option) {
   return new Promise((res) => {
     let comments = getData(token)
     comment = comments[id]
-    switch(option) {
-        case "upVote":
-            comment.voteScore = comment.voteScore + 1
-            break
-        case "downVote":
-            comment.voteScore = comment.voteScore - 1
-            break
-        default:
-            console.log(`comments.vote received incorrect parameter: ${option}`)
+    switch (option) {
+      case "upVote":
+        comment.voteScore = comment.voteScore + 1
+        break
+      case "downVote":
+        comment.voteScore = comment.voteScore - 1
+        break
+      default:
+        console.log(`comments.vote received incorrect parameter: ${option}`)
     }
     res(comment)
   })
 }
 
-function disableByParent (token, post) {
-    return new Promise((res) => {
-        let comments = getData(token)
-        keys = Object.keys(comments)
-        filtered_keys = keys.filter(key => comments[key].parentId === post.id)
-        filtered_keys.forEach(key => comments[key].parentDeleted = true)
-        res(post)
-    })
+function disableByParent(token, post) {
+  return new Promise((res) => {
+    let comments = getData(token)
+    keys = Object.keys(comments)
+    filtered_keys = keys.filter(key => comments[key].parentId === post.id)
+    filtered_keys.forEach(key => comments[key].parentDeleted = true)
+    res(post)
+  })
 }
 
-function disable (token, id) {
-    return new Promise((res) => {
-      let comments = getData(token)
-      comments[id].deleted = true
-      res(comments[id])
-    })
+function disable(token, id) {
+  return new Promise((res) => {
+    let comments = getData(token)
+    comments[id].deleted = true
+    res(comments[id])
+  })
 }
 
-function edit (token, id, comment) {
-    return new Promise((res) => {
-        let comments = getData(token)
-        for (prop in comment) {
-            comments[id][prop] = comment[prop]
-        }
-        res(comments[id])
-    })
+function edit(token, id, comment) {
+  return new Promise((res) => {
+    let comments = getData(token)
+    for (prop in comment) {
+      comments[id][prop] = comment[prop]
+    }
+    res(comments[id])
+  })
 }
 
 module.exports = {
