+++ server/api/posts.js
@@ -11,7 +11,7 @@ const defaultData = {
     author: 'thingtwo',
     category: 'react',
     voteScore: 6,
-    deleted: false 
+    deleted: false
   },
   "6ni6ok3ym7mf1p33lnez": {
     id: '6ni6ok3ym7mf1p33lnez',
@@ -25,7 +25,7 @@ const defaultData = {
   }
 }
 
-function getData (token) {
+function getData(token) {
   let data = db[token]
   if (data == null) {
     data = db[token] = clone(defaultData)
@@ -33,7 +33,7 @@ function getData (token) {
   return data
 }
 
-function getByCategory (token, category) {
+function getByCategory(token, category) {
   return new Promise((res) => {
     let posts = getData(token)
     let keys = Object.keys(posts)
@@ -42,18 +42,18 @@ function getByCategory (token, category) {
   })
 }
 
-function get (token, id) {
+function get(token, id) {
   return new Promise((res) => {
     const posts = getData(token)
     res(
-      posts[id].deleted 
+      posts[id].deleted
         ? {}
         : posts[id]
     )
   })
 }
 
-function getAll (token) {
+function getAll(token) {
   return new Promise((res) => {
     const posts = getData(token)
     let keys = Object.keys(posts)
@@ -62,10 +62,10 @@ function getAll (token) {
   })
 }
 
-function add (token, post) {
+function add(token, post) {
   return new Promise((res) => {
     let posts = getData(token)
-    
+
     posts[post.id] = {
       id: post.id,
       timestamp: post.timestamp,
@@ -76,45 +76,45 @@ function add (token, post) {
       voteScore: 1,
       deleted: false
     }
-     
+
     res(posts[post.id])
   })
 }
 
-function vote (token, id, option) {
+function vote(token, id, option) {
   return new Promise((res) => {
     let posts = getData(token)
     post = posts[id]
-    switch(option) {
-        case "upVote":
-            post.voteScore = post.voteScore + 1
-            break
-        case "downVote":
-            post.voteScore = post.voteScore - 1
-            break
-        default:
-            console.log(`posts.vote received incorrect parameter: ${option}`)
+    switch (option) {
+      case "upVote":
+        post.voteScore = post.voteScore + 1
+        break
+      case "downVote":
+        post.voteScore = post.voteScore - 1
+        break
+      default:
+        console.log(`posts.vote received incorrect parameter: ${option}`)
     }
     res(post)
   })
 }
 
-function disable (token, id) {
-    return new Promise((res) => {
-      let posts = getData(token)
-      posts[id].deleted = true
-      res(posts[id])
-    })
+function disable(token, id) {
+  return new Promise((res) => {
+    let posts = getData(token)
+    posts[id].deleted = true
+    res(posts[id])
+  })
 }
 
-function edit (token, id, post) {
-    return new Promise((res) => {
-        let posts = getData(token)
-        for (prop in post) {
-            posts[id][prop] = post[prop]
-        }
-        res(posts[id])
-    })
+function edit(token, id, post) {
+  return new Promise((res) => {
+    let posts = getData(token)
+    for (prop in post) {
+      posts[id][prop] = post[prop]
+    }
+    res(posts[id])
+  })
 }
 
 module.exports = {
